% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ask.R
\name{ask}
\alias{ask}
\title{Returns TRUE or FALSE from Y/n or y/N prompt.}
\usage{
ask(question, default = TRUE)
}
\arguments{
\item{question}{input Yes / No question}

\item{default}{Optionally set the default response value}
}
\value{
boolean response from Yes / No
}
\description{
Returns TRUE or FALSE from Y/n or y/N prompt.
}
\examples{
ask("Do you like coding?")
}
