% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ask.R
\name{ask_confirm}
\alias{ask_confirm}
\title{Utility ask function that confirms and returns the response}
\usage{
ask_confirm(question)
}
\arguments{
\item{question}{input question}
}
\value{
response
}
\description{
A function to ask a question and return the response after confirming
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
ask_confirm("Where do you live?")
}
}
