% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ask.R
\name{askyn_confirm}
\alias{askyn_confirm}
\title{Utility ask function that confirms and returns TRUE / FALSE}
\usage{
askyn_confirm(question, default = TRUE)
}
\arguments{
\item{question}{input Yes / No question}

\item{default}{Optionally set the default response value}
}
\value{
Y/y or N/n
}
\description{
A function to ask a yes/no question and return Y/N after confirming
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
ask_yn_confirm("Do you like coding?")
}
}
