% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{cheat}
\alias{cheat}
\title{Open cheatsheet}
\usage{
cheat(sheet)
}
\arguments{
\item{sheet}{Bare, unquoted name of the cheat sheet, if known. Partial matches work as well. If no argument is provided a list is presented.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Shortcut to open a cheatsheet
}
