% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{common_vars}
\alias{common_vars}
\title{common_vars}
\usage{
common_vars(...)
}
\arguments{
\item{...}{Bare, unquoted tibble object names.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Shows the variable names that are in common between two or more tibbles.
Considering deprecating this as janitor::compare_df_cols() does something similar.
}
