% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{convert_winpath}
\alias{convert_winpath}
\title{Utility Windows function}
\usage{
convert_winpath(x)
}
\arguments{
\item{question}{input Yes / No question}

\item{default}{Optionally set the default response value}
}
\value{
boolean response from Yes / No
}
\description{
A function to convert a path generated by "copy as path"
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
convert_winpath("testpath")
}
}
