% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_log.R
\name{copy_file_wlog}
\alias{copy_file_wlog}
\title{Copy file and log where it was copied from}
\usage{
copy_file_wlog(path, new_path, overwrite = FALSE)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{new_path}{A character vector of paths to the new locations.}

\item{overwrite}{Overwrite files if they exist. Default is FALSE. If this is FALSE and the file exists an error will be thrown.}
}
\value{
starting or adding a file_log.csv at project-level, as well as copying file as expected
}
\description{
A wrapper for fs::file_copy() that adds a project-level log with pertinent info into file_log.csv
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Copy file with log
}
\examples{
\dontrun{
.old_wd <- setwd(tempdir())
fs::file_create("foo")
copy_file_wlog("foo", "bar")
}
}
