% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Creates a new R Project and populates with Integral standards}
\usage{
create_project(
  create_dirs = ask("Would you like to create input, output and QA folders?"),
  create_rproj = ask("Would you like to create an Rproj file?"),
  create_rscript = ask("Would you like to creat an R script file?"),
  create_rmarkdown = ask("Would you like to create an Rmarkdown file?"),
  create_qa = ask("Would you like to create a QA file?"),
  create_git = ask("Would you like to create a git repository?"),
  create_renv = ask("Would you like to create an renv?"),
  swtich_proj = ask("Would you like to switch to the newly created project?")
)
}
\description{
Creates a new R Project and populates with Integral standards
}
