% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-path.R
\name{determine_path}
\alias{determine_path}
\title{determine_path}
\usage{
determine_path()
}
\value{
path to file
}
\description{
Function to find file path
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
determine_path()
}
}
