% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{find_file}
\alias{find_file}
\title{Find a new file interactively
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}}
\usage{
find_file(ext, startdir = rstudioapi::getActiveProject())
}
\arguments{
\item{ext}{file extension to check for. Can be R, Rmd etc}

\item{startdir}{Optional start directory}
}
\description{
Finds a new file from user inout
}
\examples{
\dontrun{
find_file("Rmd")
}

}
