% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_custom_formatting.R
\name{get_breaks}
\alias{get_breaks}
\title{Get list of breaks to be used as drows argument in \code{to_fff}}
\usage{
get_breaks(dat, col_num = 1, rm_grps = 1)
}
\arguments{
\item{dat}{Unformatted dataframe containing crosstabbed data.}

\item{col_num}{Numeric value specifying column number to use for identifying breaks}

\item{rem_grps}{Numeric vector specifying group indices to remove (defaults to 1 for headers)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
