% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_system}
\alias{get_system}
\title{Determine which system R is running on}
\usage{
get_system()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Returns one of: "citrix", "linux", "local" depending on which integral environment R is being run on.
}
