% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_search_helpers.R
\name{google}
\alias{google}
\title{Search Google for R related results}
\usage{
google(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Stackoverflow directly from R console.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The required term "R" is automatically included.
Note: Prevents results from sites that show the same info as the help documentation within R.
}
\examples{
google("date conversion")
}
\keyword{stackoverflow}
\keyword{web}
\keyword{workflow}
