% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_search_helpers.R
\name{google_last_error}
\alias{google_last_error}
\title{Search Google for your most recent error message.}
\usage{
google_last_error()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
