% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_custom_functions.R
\name{huck}
\alias{huck}
\title{Remove a named list element}
\usage{
huck(.x, element)
}
\arguments{
\item{element_name}{Name of the element to remove}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Opposite of \code{purrr::pluck()}.  Returns the original list without the element.
}
