% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ic_copy_file}
\alias{ic_copy_file}
\title{Start a new file with header
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ic_copy_file(filepath, rfile_path, open = F)
}
\arguments{
\item{filepath}{File to be created.}

\item{rfile_path}{R file to source}

\item{open}{Should the file be open (True/False)}
}
\description{
Creates a new file with header. Function is not exported.
}
\examples{
\dontrun{
ic_new_file("myfile.R", fs::path_package(
  "integral",
  "templates/example_project/rfile_w_header.R"
))
}
}
