% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ic_new_git}
\alias{ic_new_git}
\title{Start a new git repository with Integral gitignore
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ic_new_git(dirpath)
}
\arguments{
\item{dirpath}{File to be created.}
}
\description{
Creates a new git repository at the given path
}
\examples{
\dontrun{
ic_new_git("some directory path")
}

}
