% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_proj.R
\name{ic_new_proj}
\alias{ic_new_proj}
\title{Create a new project and R files following Integral standards}
\usage{
ic_new_proj(
  path = "",
  create_dirs = T,
  create_rproj = T,
  create_git = ask("Would you like to create a git repository?"),
  switch_proj = ask("Would you like to switch to the newly created project?")
)
}
\arguments{
\item{path}{Path to new project.  Defaults to user input during function.}

\item{create_dirs}{Y/n whether to create input, output and QA folders}

\item{create_rproj}{Y/n whether to create a .Rproj file}

\item{create_git}{Y/n whether to create a git repository}

\item{switch_projf}{Y/n whether to switch to newly created project
Creates a new R Project and populates with Integral standards}
}
\value{
path to new project
}
\description{
Function to create a new R project and files following integral
standards
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
new_proj()
}
}
