% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ic_new_python_file}
\alias{ic_new_python_file}
\title{Start a new Python script file with Integral's header
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ic_new_python_file(filepath = "", open = TRUE)
}
\arguments{
\item{filepath}{File to be created.}

\item{open}{Logical. Defaults to \code{TRUE}}
}
\description{
Creates a new script with header
}
\examples{
\dontrun{
ic_new_python_file("myfile.py")
}
}
