% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ic_new_r_file}
\alias{ic_new_r_file}
\title{Start a new R script with header}
\usage{
ic_new_r_file(filepath)
}
\arguments{
\item{filepath}{File to be created. If path is not included, it will be created in the current working directory.}
}
\description{
Creates a new R script with header
}
\examples{
\dontrun{
ic_new_r_file("myfile.R")
}
}
