% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ic_new_script}
\alias{ic_new_script}
\alias{ic_new_r_file}
\alias{ic_new_python_file}
\title{Start a new script file with Integral's header
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ic_new_script(filepath, open = F)

ic_new_r_file(filepath = "", open = F)

ic_new_python_file(filepath = "", open = F)
}
\arguments{
\item{filepath}{File to be created.}

\item{open}{Open the file in the default editor (TRUE/FALSE)}
}
\description{
Creates a new script with header
}
\examples{
\dontrun{
ic_new_script("myfile.R")
ic_new_r_file("myfile.R")
ic_new_python_file("myfile.py")
}
}
