% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptfig.R
\name{ic_pptfig}
\alias{ic_pptfig}
\title{Function to place R generated figures into a PowerPoint Template}
\usage{
ic_pptfig(
  my_pres = NA,
  size = "letter",
  orientation = "L",
  fig = last_plot(),
  section = "",
  fignum,
  note = "",
  author,
  doctitle1,
  doctitle2 = "",
  doctitle3 = "",
  draft = "DRAFT",
  pc = "Privileged and Confidential",
  showpath = T
)
}
\arguments{
\item{my_pres}{template pptx, defaults to read from read_pptx_template function}

\item{size}{Presentation size: letter (default) or ledger.}

\item{orientation}{Presentation orientation: landscape (L, default) or Portrait (P)}

\item{fig}{graphics object that is compatible with officer, defaults to last last_plot()}

\item{section}{Optional section number}

\item{fignum}{Figure number after the section-figure number}

\item{note}{Optional figure notes text to include under figure}

\item{author}{The script author}

\item{doctitle1}{Document Title 1}

\item{doctitle2}{Optional Document Title 2}

\item{doctitle3}{Optional Document Title 3}

\item{draft}{Bold red text to be included next to figure caption, "DRAFT" is default}

\item{pc}{Privileged and Confidential label, "Privileged and Confidential" is default}

\item{showpath}{Show script path on page. Default is True}
}
\value{
PowerPoint presentation with added figure
Note: function dml() creates an editable figure in PowerPoint
fig <- dml(ggobj = p) #for ggplot object
fig <- dml({grid.arrange(p1, p2, p3, p4, nrow = 4)})
}
\description{
Function to place R generated figures into a PowerPoint Template
}
\examples{
#' my_pres <- ic_pptfig( fig = ggplot2::ggplot(), fignum = "1", author = "J. Doe", doctitle1 = "example figure")

# Print presentation in calling script
# print(my_pres, "Output_File-Name.pptx") #must end in .pptx
}
