% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{ic_scale_si_unit}
\alias{ic_scale_si_unit}
\title{Scale format to International System of Units (k, M, )}
\usage{
ic_scale_si_unit(sep = " ", ...)
}
\arguments{
\item{sep}{Seperator to use between number and unit (defaults to " ")}

\item{...}{Passed by \code{ggplot2::scale_*_continuous}}
}
\value{
Used with scale_*_continuous, returns formatted axis labels
}
\description{
#' \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Format a vector of numeric values according to the International System of Units.

(ie: 1,000 becomes "1 k", 2,500 becomes "2.5 k") See: http://en.wikipedia.org/wiki/SI_prefix
}
\examples{
\dontrun{
ggplot(diamonds, aes(x = price)) + geom_density() + scale_x_continuous(labels = ic_scale_si_unit())
}
}
