% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ic_update}
\alias{ic_update}
\title{Update the integral package}
\usage{
ic_update()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Automatically unloads, updates, and reloads the integral package.
}
