% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visdat_custom_functions.R
\name{ic_visdat_grouped}
\alias{ic_visdat_grouped}
\title{vis_dat for grouped data}
\usage{
ic_visdat_grouped(.data, ..., method = "vis_dat", .sample_frac = "auto")
}
\arguments{
\item{...}{bare, unquoted column(s) to use as the index to group by. Alternatively will accept a grouped df.}

\item{method}{Which visdat function to use. One of "vis_dat", "vis_miss", or "vis_value".  Accepts shorthand "dat", "val", and "miss".}

\item{.sample_frac}{Percent of observations to sample from each year.  Default "auto" samples down to 100,000 rows, split evenly between groups for vis_dat. For vis_miss and vis_value, "auto" uses all data.}
}
\value{
visdat plot separated by grouping variable.
}
\description{
Produce a vis_dat plot for ipeds data split by year with optional sampling.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

Note that parallel processing is built in if a \code{future::plan()} is set
}
\examples{
\dontrun{
diamonds \%>\% ic_visdat_group(facet_group = cut)
}
}
