% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_log.R
\name{ic_write_csv_wlog}
\alias{ic_write_csv_wlog}
\title{Add to save log when csv is written}
\usage{
ic_write_csv_wlog(x, file, ...)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{file}{File or connection to write to.}

\item{...}{Additional arguments from readr::write_csv()}
}
\value{
starting or adding a file_log.csv at project-level, as well as saving csv as expected
}
\description{
A wrapper for write_csv that adds a project-level log with pertinent info into file_log.csv
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Write csv with log
}
\examples{
\dontrun{
df <- data.frame(fishes <- c("red", "white", "blue"))
ic_write_csv_wlog(df, tempfile())
}
}
