% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_citrix}
\alias{is_citrix}
\title{Check whether R is running on Citrix}
\usage{
is_citrix()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Returns TRUE if the current system is Citrix.
}
