% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-columns.R
\name{letnum}
\alias{letnum}
\title{convert letter to column number}
\usage{
letnum(x)
}
\arguments{
\item{x}{A letter to convert to a column number starting with 1.}
}
\value{
Column number as integer
}
\description{
Converts an upper / lowercase letter into the resulting column
number. Starts at 1. Limited to current Excel maximum of XFD (16,384 columns)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
letnum("A")

History
2022-05 Eben Pendleton written.
}
