% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_proj.R
\name{new_proj}
\alias{new_proj}
\title{Create a new project and R files following Integral standards}
\usage{
new_proj(
  create_dirs = ask("Would you like to create input, output and QA folders?"),
  create_rproj = ask("Would you like to create an Rproj file?"),
  create_rscript = ask("Would you like to create an R script file?"),
  create_rmarkdown = ask("Would you like to create an Rmarkdown file?", default = F),
  create_qa = ask("Would you like to create a QA file?"),
  create_git = ask("Would you like to create a git repository?"),
  create_renv = ask("Would you like to create an renv?"),
  switch_proj = ask("Would you like to switch to the newly created project?")
)
}
\arguments{
\item{create_dirs}{Y/n whether to create input, output and QA folders}

\item{create_rproj}{Y/n whether to create a .Rproj file}

\item{create_rscript}{Y/n whether to create a R script file}

\item{create_rmarkdown}{Y/n whether to create a R Markdown file}

\item{create_qa}{Y/n whether to create a QA file}

\item{create_git}{Y/n whether to create a git repository}

\item{create_renv}{Y/n whether to create a renv

Creates a new R Project and populates with Integral standards}
}
\value{
path to new project
}
\description{
Function to create a new R project and files following integral
standards
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
new_proj()
}
}
