% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptfig.R
\name{pptfig}
\alias{pptfig}
\title{Function to place R generated figures into a PowerPoint Template}
\usage{
pptfig(
  my_pres,
  size,
  orientation,
  fig,
  section,
  fignum,
  note,
  author,
  doctitle1,
  doctitle2 = "",
  doctitle3 = "",
  draft = "DRAFT",
  pc = "Privileged and Confidential",
  showpath = T
)
}
\arguments{
\item{my_pres}{from read_pptx_template function}

\item{size}{Presentation size (letter or ledger)}

\item{orientation}{Presentation orientation landscape (L) or Portrait (P)}

\item{fig}{graphics object that is compactiable with officer}

\item{section}{section number X-}

\item{fignum}{figure number after the section-figure number}

\item{note}{Figure notes text to include under figure}

\item{author}{The script author}

\item{doctitle1}{Document Title 1}

\item{doctitle2}{Optional Document Title 2}

\item{doctitle3}{Optional Document Title 3}

\item{draft}{Optional Bold red text to be included next to figure caption (ex. "DRAFT")}

\item{pc}{Optional Privileged and Confidential label}

\item{showpath}{Optional Show script path on page. Default is True}
}
\value{
PowerPoint presentation with added figure
Note: function dml() creates an editable figure in PowerPoint
fig <- dml(ggobj = p) #for ggplot object
fig <- dml({grid.arrange(p1, p2, p3, p4, nrow = 4)})
}
\description{
Function to place R generated figures into a PowerPoint Template
}
\examples{
size <- c("letter")
orientation <- c("L")
fig <- ggplot2::ggplot()
section <- c("X")
fignum <- c("1")
note <- "Figure Note"
author <- "Eben Pendleton"
doctitle1 <- "Document Title"
my_pres <- read_pptx_template(size, orientation)
my_pres <- pptfig(my_pres, size, orientation, fig, section, fignum, note, author,
  doctitle1,
  doctitle2 = "", doctitle3 = "",
  draft = "DRAFT"
)

# Print presentation in calling script
# print(my_pres, "Output_File-Name.pptx") #must end in .pptx
}
