% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa.R
\name{qa}
\alias{qa}
\title{Automated QA form population}
\usage{
qa(filepath = "youfile.R")
}
\arguments{
\item{filepath}{The file to QA. Can include either an absolute path or a relative path (including "~" home references). If omitted, a file selection dialog box will appear.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Scrape an .R or .Rmd file for QA tags, and populate the Integral QA sheet for scripted analyses.

To use, tag QA review items by commenting:

QA: Review comment / request
}
\section{Function Process}{


The function will perform the following actions:

\itemize{
\item Scrape the file for section headers and '# QA:' comment tags
\item Assign unique ID's to QA tags in the code (the file will be modified) if they do not have one.
\item Create a QA subdirectory and a QA excel workbook if one does not exist
\item Add a code review spreadsheet for the file if one does not exist; update the spreadsheet for the file if one already exists
}
}

