% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{quote_values}
\alias{quote_values}
\title{Convert raw comma separated values into a vector}
\usage{
quote_values(...)
}
\arguments{
\item{...}{any number of names separated by commas}
}
\value{
character string vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Converts a series of comma-separated values into vector input format for R.
(Note, this is identical to Hmisc::Cs).
}
\examples{
\dontrun{
quote_values(a, b, c, d, e, f)
}
}
