% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptfig.R
\name{read_pptx_template}
\alias{read_pptx_template}
\title{Function to place R generated figures into a PowerPoint Template}
\usage{
read_pptx_template(size, orientation)
}
\arguments{
\item{size}{Presentation size (letter or ledger)}

\item{orientation}{Presentation orientation landscape (L) or Portrait (P)}
}
\value{
PowerPoint presentation template
}
\description{
Function to place R generated figures into a PowerPoint Template
}
\examples{
size <- c("letter")
orientation <- c("L")
my_pres <- read_pptx_template(size, orientation)
}
