% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_custom_formatting.R
\name{read_wb}
\alias{read_wb}
\title{Read Excel workbook(s) and return nested list with all workbooks and worksheets}
\usage{
read_wb(fpath, fname = NULL, sheet = NULL)
}
\arguments{
\item{fpath}{File path containing Excel workbooks}

\item{fname}{File name(s) of workbooks}

\item{sheet}{Sheet index to extract}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
