% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{slice_ht}
\alias{slice_ht}
\title{Slice head and tail}
\usage{
slice_ht(.data, n, prop)
}
\arguments{
\item{.data}{A data frame, or data frame extension (e.g. a tibble).}

\item{n}{The number of rows to slice from both head and tail (e.g. n = 5 means 5 from head and 5 from tail). Defaults to 5.}

\item{prop}{asdf}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This is a modified version of dplyr::slice_* that allows slicing n rows from both head and tail.
}
