% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_search_helpers.R
\name{so_last_error}
\alias{so_last_error}
\title{Search Stack Overflow for your most recent error message.}
\usage{
so_last_error()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
