% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_and_search_helpers.R
\name{stackoverflow}
\alias{stackoverflow}
\alias{so}
\title{Search Stackoverflow}
\usage{
stackoverflow(search_terms)

so(search_terms)
}
\arguments{
\item{search_terms}{Search terms encapsulated in " ".}
}
\description{
Improve your workflow by searching Stackoverflow directly from R console.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
stackoverflow("r date conversion")
so("r ggplot2 geom_smooth()")
}
\keyword{stackoverflow}
\keyword{web}
\keyword{workflow}
