% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{state_locations}
\alias{state_locations}
\title{Identify the location of states to facilitate direction of popups.}
\usage{
state_locations(albers = TRUE)
}
\arguments{
\item{albers}{specify whether the map being used is shifted to albers. Defaults to TRUE.}
}
\value{
Table with relative location of each state
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Returns a table of state names, state abbreviations, and the location of each state.
}
