% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_between}
\alias{str_between}
\title{Select string between two strings}
\usage{
str_between(string, pattern1, pattern2)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible to one.}

\item{pattern1}{Pattern that will act as starting anchor in extraction, and be excluded. The default interpretation is a regular expression, as described in stringi::stringi-search-regex. Control options with regex().}

\item{pattern2}{Pattern that will act as ending anchor in extraction, and be excluded.}
}
\value{
string or vector of strings
}
\description{
A stringr-like function to pull out a string between two string anchors
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# str_between("apple", "ap", "e")
}
