% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_squish_line_breaks}
\alias{str_squish_line_breaks}
\title{Trim strings with repeated line breaks}
\usage{
str_squish_line_breaks(string)
}
\arguments{
\item{string}{String to squish}
}
\value{
String with repeeated line returns and whitespaces removed
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Removes line breaks  (\\n and \\r) from the start and end of a string and
repeated line breaks inside a string.

This is an analogue of stringr::str_squish(), which trims white space from
the start and end of a string and removes repeated whitespace inside a
string. However stringr::str_squish() will replace line breaks inside a
string with a space, which is not always desired.

Note that if there is a repeat of two different line break characters, i.e. a newline (\\n)
and line return (\\r), the first occurrence will be kept.

See: https://github.com/tidyverse/stringr/issues/477
}
\examples{
stringr::str_squish("hello\nworld")
stringr::str_squish("hello\n\nworld")
str_squish_line_breaks("hello\nworld")
str_squish_line_breaks("hello\n\nworld")
}
