% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_custom_functions.R
\name{tp}
\alias{tp}
\title{Tibble Preview}
\usage{
tp(.data = .Last.value, rows = 10)
}
\arguments{
\item{.data}{a tibble or data frame (default: .Last.value)}

\item{rows}{number of rows to print (default: 10)}
}
\value{
A preview of a tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Show a sample of all tibble data without hiding columns.
}
