% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{wc}
\alias{wc}
\title{Write Clip shortcut}
\usage{
wc(x = .Last.value)
}
\arguments{
\item{x}{any text you'd like copied to clipboard}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Copies an object to the clipboard. Defaults to .Last.value#'
}
