% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xview.R
\name{xview}
\alias{xview}
\title{Show a table in Excel}
\usage{
xview(.data, add_rownames = TRUE)
}
\arguments{
\item{.data}{A table.}

\item{add_rownames}{Logical. Convert rownames to column? Default: TRUE}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Opens the table in a temporary file in excel.  Currently only works on Mac OS.
}
\examples{
\dontrun{
diamonds \%>\% xview()
}
}
