% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ask.R, R/utils-path.R
\name{ask}
\alias{ask}
\alias{convert_winpath}
\title{Utility ask function}
\usage{
ask(question, default = TRUE)

convert_winpath(x)
}
\arguments{
\item{question}{input Yes / No question}

\item{default}{Optionally set the default response value}
}
\value{
boolean response from Yes / No

boolean response from Yes / No
}
\description{
A function to ask a yes/no question and return TRUE/FALSE
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A function to ask a yes/no question and return TRUE/FALSE
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
ask("Do you like coding?")
}
\dontrun{
convert_winpath(<path>)
}
}
