% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{cheat}
\alias{cheat}
\title{Open cheatsheet}
\usage{
cheat(sheet)
}
\arguments{
\item{sheet}{Bare, unquoted name of the cheat sheet, if known. Partial matches work as well. If NULL (default) a list is presented.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Shortcut to open a cheatsheet
}
