% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purrr_custom_functions.R
\name{huck}
\alias{huck}
\title{Remove a named list element}
\usage{
huck(.data, ...)
}
\arguments{
\item{.data}{a named list}

\item{...}{Unquoted variable names to search for duplicates. This takes a tidyselect specification (starts_with, contains, ends_with, etc).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Opposite of \code{purrr::pluck()}.  Returns the original list without the element.
}
\examples{
obj1 <- list("a", list(1, elt = "foo"))
obj2 <- list("b", list(2, elt = "bar"))
x <- list(a = obj1, b = obj2)
x \%>\% huck(a)
x \%>\% huck(starts_with("a"))
}
