% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ic_news}
\alias{ic_news}
\title{Show news for new version}
\usage{
ic_news(all = FALSE)
}
\arguments{
\item{all}{Logical. Show all previous news in viewer. Defaults to FALSE.}
}
\description{
\lifecycle{experimental}
Shows the news for the package
}
