% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{ig_new_rmd_file}
\alias{ig_new_rmd_file}
\title{Start a new RMarkdown document with Integral header}
\usage{
ig_new_rmd_file(file)
}
\arguments{
\item{filepath}{File to be created. If path isn't provided, working directory is used.}
}
\description{
Drafts a new RMarkdown document with Integral header
}
\examples{
\dontrun{
ig_new_rmd_file("myrmd.Rmd")
}
}
