% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral-package.R
\docType{package}
\name{integral-package}
\alias{integral}
\alias{integral-package}
\title{integral: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://integralenvision.github.io/integral}
}

}
\author{
\strong{Maintainer}: Jonathan Zadra \email{jzadra@integral-corp.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
