% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral-package.R
\docType{package}
\name{integral-package}
\alias{integral}
\alias{integral-package}
\title{integral: Integral Consulting Public Package}
\description{
This package contains convenience functions used by Integral Consulting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://integralenvision.github.io/integral}
}

}
\author{
\strong{Maintainer}: Jonathan Zadra \email{jzadra@integral-corp.com}

Authors:
\itemize{
  \item Katherine Heal \email{kheal@integral-corp.com}
}

}
\keyword{internal}
